Add distribution code for Filter (more) problem of Week 4 problem set
The "more" part in the name of the problem indicates that this is the
"more comfortable" version of the Filter problem in the problem set.
The difference in the two versions is that this "more comfortable"
version includes an edge detection filter, whereas the
"less comfortable" version includes a sepia filter instead.