Add The Girl Who Leapt Through Time web page to the website
Added the content, styles, and interactivity associated with the new
web page for the movie The Girl Who Leapt Through Time. Additionally,
the content format, styles, and interactivity for this web page should
be mostly, if not yet completely, generalized so that the future
web pages for the other movies can be easily added and will retain a
cohesive format, style, and interactivity.