Add Summer Wars web page to the website
Added content associated with the new web page for the movie Summer
Wars. The styles and interactivity were mostly left unchanged, with
the exception of adding new styles and interactivity for buttons with
a dark outline. Additionally, the other pages in the website were
modified to include a link to this new web page that was added.