Add Mirai web page to the website
Added a new CSS declaration for the styles of this web page, changing
the background color to white and the text color to black, and changed
the button outlines to dark in the HTML file. Additionally, the other
web pages in the website were modified to include a link to this web
page.