Implement sell route
The user is now able to sell shares from any stock that they have
previously bought.

Additionally, the purchases table in the database was modified to now
be a transactions table, which keeps track of both shares bought and
sold.

A new table (user_portfolios) was also created in the database to keep
track of the stocks and the amount of shares of each stock currently
owned by any user. This new table is updated alongside the transactions
table any time a user buys or sells shares of a stock, and it is used
to display a user's portfolio in their homepage.