Implement history route
Users will now be able to see a table with the history of all of the
transactions that they have made. The table displays the stocks that the
user bought or sold, the number of shares bought or sold, the price of
the stock at the moment of the transaction, the total value of that
transaction, and the date in which the transaction took place.